# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IAA 投放运营中台 API",
    "description": "IAA 投放运营中台 Java 业务服务契约；示例数据均为虚构。模型暂由 FastAPI 导出，x-implementation 表示 Java 状态；x-reference-implementation 保留 Python 对照服务的实际状态。",
    "version": "1.0.0-rc.1"
  },
  "paths": {
...
    "/api/v1/filter-options": {
      "get": {
        "tags": [
          "看盘分析"
        ],
        "summary": "筛选条可选值",
        "description": "返回日期范围内、当前账号数据范围可见的媒体、产品、代理、账户与运营。",
        "operationId": "getFilterOptions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "可选值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterOptions"
                },
                "examples": {
                  "demo": {
                    "summary": "运营账号可见范围",
                    "value": {
                      "media": [
                        {
                          "value": "vivo",
                          "label": "vivo"
                        },
                        {
                          "value": "oppo",
                          "label": "OPPO"
                        }
                      ],
                      "products": [
                        {
                          "value": "晴空天气",
                          "label": "晴空天气"
                        },
                        {
                          "value": "脑力答题王",
                          "label": "脑力答题王"
                        }
                      ],
                      "agencies": [
                        {
                          "value": "星河代理",
                          "label": "星河代理"
                        },
                        {
                          "value": "蓝鲸代理",
                          "label": "蓝鲸代理"
                        }
                      ],
                      "accounts": [
                        {
                          "value": "vivo-demo-003",
                          "label": "vivo-demo-003"
                        },
                        {
                          "value": "vivo-demo-017",
                          "label": "vivo-demo-017"
                        }
                      ],
                      "operators": [
                        {
                          "value": "运营甲",
                          "label": "运营甲"
                        },
                        {
                          "value": "运营乙",
                          "label": "运营乙"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无权限，或超出账号的数据范围",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-requirements": [
          "IAA-REQ-002"
        ],
        "x-screens": [
          "app-shell"
        ],
        "x-fields": [
          "stat_date",
          "media",
          "product",
          "agency",
          "account",
          "operator"
        ],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "implemented"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/reporting/web/ReportController.java"
      }
    },
    "/api/v1/reports/aggregate": {
      "post": {
        "tags": [
          "看盘分析"
        ],
        "summary": "聚合查询",
        "description": "按所选维度汇总；派生指标由汇总后的基础指标重算，合计行同理。结果受账号数据范围约束。",
        "operationId": "queryAggregate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportQuery"
              },
              "examples": {
                "demo": {
                  "summary": "vivo 按产品聚合",
                  "value": {
                    "date_from": "2026-09-14",
                    "date_to": "2026-09-15",
                    "filters": {
                      "media": [
                        "vivo"
                      ]
                    },
                    "metrics": [
                      "cost",
                      "revenue",
                      "roi",
                      "cpc",
                      "click_arpu"
                    ],
                    "sort": [
                      {
                        "field": "cost"
                      }
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "聚合结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResult"
                },
                "examples": {
                  "demo": {
                    "summary": "按产品聚合",
                    "value": {
                      "columns": [
                        {
                          "key": "product",
                          "label": "产品",
                          "kind": "dimension"
                        },
                        {
                          "key": "cost",
                          "label": "消耗",
                          "kind": "base",
                          "unit": "money",
                          "precision": 2
                        },
                        {
                          "key": "revenue",
                          "label": "预估收益",
                          "kind": "base",
                          "unit": "money",
                          "precision": 2
                        },
                        {
                          "key": "roi",
                          "label": "ROI",
                          "kind": "derived",
                          "unit": "ratio",
                          "precision": 3
                        },
                        {
                          "key": "cpc",
                          "label": "CPC",
                          "kind": "derived",
                          "unit": "per_unit",
                          "precision": 4
                        },
                        {
                          "key": "click_arpu",
                          "label": "点击 ARPU",
                          "kind": "derived",
                          "unit": "per_unit",
                          "precision": 4
                        }
                      ],
                      "rows": [
                        {
                          "product": "晴空天气",
                          "cost": 12840.5,
                          "revenue": 14125.3,
                          "roi": 1.1,
                          "cpc": 0.1333,
                          "click_arpu": 0.1467
                        },
                        {
                          "product": "脑力答题王",
                          "cost": 8420.0,
                          "revenue": 7910.6,
                          "roi": 0.94,
                          "cpc": 0.1401,
                          "click_arpu": 0.1316
                        }
                      ],
                      "totals": {
                        "cost": 21260.5,
                        "revenue": 22035.9,
                        "roi": 1.036,
                        "cpc": 0.1359,
                        "click_arpu": 0.1409
                      },
                      "page": 1,
                      "page_size": 50,
                      "total_rows": 2,
                      "data_as_of": "2026-09-15T23:50:00+08:00"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无权限，或超出账号的数据范围",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-requirements": [
          "IAA-REQ-002"
        ],
        "x-screens": [
          "report-aggregate"
        ],
        "x-fields": [
          "stat_date",
          "media",
          "product",
          "agency",
          "account",
          "operator",
          "cost",
          "revenue",
          "clicks",
          "roi",
          "cpc",
          "click_arpu"
        ],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "implemented"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/reporting/web/ReportController.java"
      }
    },
    "/api/v1/reports/daily": {
      "post": {
        "tags": [
          "看盘分析"
        ],
        "summary": "分天明细",
        "description": "在聚合维度前固定加入日期维度，便于逐日对比；与聚合共用同一筛选与口径。",
        "operationId": "queryDaily",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportQuery"
              },
              "examples": {
                "demo": {
                  "summary": "近两天按产品",
                  "value": {
                    "date_from": "2026-09-14",
                    "date_to": "2026-09-15",
                    "filters": {
                      "media": [
                        "vivo"
                      ]
                    },
                    "group_by": [
                      "stat_date",
                      "product"
                    ],
                    "metrics": [
                      "cost",
                      "revenue",
                      "roi"
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "分天结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResult"
                },
                "examples": {
                  "demo": {
                    "summary": "按日期与产品",
                    "value": {
                      "columns": [
                        {
                          "key": "stat_date",
                          "label": "日期",
                          "kind": "dimension"
                        },
                        {
                          "key": "product",
                          "label": "产品",
                          "kind": "dimension"
                        },
                        {
                          "key": "cost",
                          "label": "消耗",
                          "kind": "base",
                          "unit": "money",
                          "precision": 2
                        },
                        {
                          "key": "revenue",
                          "label": "预估收益",
                          "kind": "base",
                          "unit": "money",
                          "precision": 2
                        },
                        {
                          "key": "roi",
                          "label": "ROI",
                          "kind": "derived",
                          "unit": "ratio",
                          "precision": 3
                        }
                      ],
                      "rows": [
                        {
                          "stat_date": "2026-09-14",
                          "product": "晴空天气",
                          "cost": 6120.4,
                          "revenue": 6905.1,
                          "roi": 1.128
                        },
                        {
                          "stat_date": "2026-09-15",
                          "product": "晴空天气",
                          "cost": 6720.1,
                          "revenue": 7220.2,
                          "roi": 1.074
                        },
                        {
                          "stat_date": "2026-09-14",
                          "product": "脑力答题王",
                          "cost": 4180.0,
                          "revenue": 3620.6,
                          "roi": 0.866
                        },
                        {
                          "stat_date": "2026-09-15",
                          "product": "脑力答题王",
                          "cost": 4240.0,
                          "revenue": 4290.0,
                          "roi": 1.012
                        }
                      ],
                      "totals": {
                        "cost": 21260.5,
                        "revenue": 22035.9,
                        "roi": 1.036
                      },
                      "page": 1,
                      "page_size": 50,
                      "total_rows": 4,
                      "data_as_of": "2026-09-15T23:50:00+08:00"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无权限，或超出账号的数据范围",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-requirements": [
          "IAA-REQ-002"
        ],
        "x-screens": [
          "report-daily"
        ],
        "x-fields": [
          "stat_date",
          "media",
          "product",
          "agency",
          "account",
          "operator",
          "cost",
          "revenue",
          "clicks",
          "roi",
          "cpc",
          "click_arpu"
        ],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "implemented"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/reporting/web/ReportController.java"
      }
    },
    "/api/v1/reports/trend": {
      "post": {
        "tags": [
          "看盘分析"
        ],
        "summary": "趋势序列",
        "description": "按天或小时输出指标序列，可按一个维度拆成最多 8 条线，其余合并为「其他」。按小时时跨度不超过 7 天。",
        "operationId": "queryTrend",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrendQuery"
              },
              "examples": {
                "demo": {
                  "summary": "按天看消耗与 ROI",
                  "value": {
                    "date_from": "2026-09-14",
                    "date_to": "2026-09-15",
                    "filters": {
                      "media": [
                        "vivo"
                      ]
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "趋势序列",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendResult"
                },
                "examples": {
                  "demo": {
                    "summary": "消耗与 ROI",
                    "value": {
                      "metrics": [
                        "cost",
                        "roi"
                      ],
                      "series": [
                        {
                          "key": "all",
                          "label": "全部",
                          "points": [
                            {
                              "bucket": "2026-09-14",
                              "values": {
                                "cost": 10300.4,
                                "roi": 1.022
                              }
                            },
                            {
                              "bucket": "2026-09-15",
                              "values": {
                                "cost": 10960.1,
                                "roi": 1.05
                              }
                            }
                          ]
                        }
                      ],
                      "data_as_of": "2026-09-15T23:50:00+08:00"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无权限，或超出账号的数据范围",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-requirements": [
          "IAA-REQ-002"
        ],
        "x-screens": [
          "report-trend"
        ],
        "x-fields": [
          "stat_date",
          "stat_hour",
          "cost",
          "roi"
        ],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "implemented"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/reporting/web/ReportController.java"
      }
    },
    "/api/v1/reports/roi-anomalies": {
      "post": {
        "tags": [
          "看盘分析"
        ],
        "summary": "ROI 异常清单",
        "description": "列出消耗达到门槛且 ROI 低于阈值的对象；阈值与门槛随请求传入并回显在 rule 中。",
        "operationId": "queryRoiAnomalies",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoiAnomalyQuery"
              },
              "examples": {
                "demo": {
                  "summary": "ROI 低于 110%",
                  "value": {
                    "date_from": "2026-09-14",
                    "date_to": "2026-09-15",
                    "filters": {
                      "media": [
                        "vivo"
                      ]
                    },
                    "group_by": [
                      "account",
                      "product"
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "异常清单",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoiAnomalyResult"
                },
                "examples": {
                  "demo": {
                    "summary": "账户维度异常",
                    "value": {
                      "rule": "ROI < 110% 且消耗 ≥ 200",
                      "items": [
                        {
                          "dimensions": {
                            "account": "vivo-demo-017",
                            "product": "脑力答题王"
                          },
                          "cost": 8420.0,
                          "revenue": 7910.6,
                          "roi": 0.94,
                          "reason": "ROI 94.0% 低于 110%"
                        }
                      ],
                      "data_as_of": "2026-09-15T23:50:00+08:00"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无权限，或超出账号的数据范围",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-requirements": [
          "IAA-REQ-002"
        ],
        "x-screens": [
          "roi-anomalies"
        ],
        "x-fields": [
          "account",
          "product",
          "cost",
          "revenue",
          "roi"
        ],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "implemented"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/reporting/web/ReportController.java"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "code": {
            "type": "string",
            "enum": [
              "AUTH_REQUIRED",
              "INVALID_CREDENTIALS",
              "PASSWORD_CHANGE_REQUIRED",
              "FORBIDDEN",
              "NOT_FOUND",
              "METHOD_NOT_ALLOWED",
              "CONFLICT",
              "PAYLOAD_TOO_LARGE",
              "VALIDATION_FAILED",
              "RESULT_TOO_LARGE",
              "INTERNAL_ERROR",
              "NOT_IMPLEMENTED",
              "HTTP_ERROR"
            ],
            "title": "Code",
            "x-code-descriptions": {
              "AUTH_REQUIRED": "未登录、令牌缺失或已失效，需要重新登录。",
              "CONFLICT": "对象已被他人修改（revision 不一致）或唯一键冲突，刷新后重试。",
              "FORBIDDEN": "没有该操作的权限，或请求超出账号的数据范围。",
              "HTTP_ERROR": "其他 HTTP 错误，以 status 为准。",
              "INTERNAL_ERROR": "服务端异常，已记录，可稍后重试。",
              "INVALID_CREDENTIALS": "账号或口令错误，或登录暂被限制；不区分账号不存在、口令错误、账号停用与锁定。",
              "METHOD_NOT_ALLOWED": "路径存在，但不支持该 HTTP 方法。",
              "NOT_FOUND": "对象或路径不存在。",
              "NOT_IMPLEMENTED": "契约已确定，接口尚未实现。",
              "PASSWORD_CHANGE_REQUIRED": "正在使用临时口令，修改口令前只能访问当前账号与改密接口。",
              "PAYLOAD_TOO_LARGE": "上传文件超过大小限制。",
              "RESULT_TOO_LARGE": "分组数或导出行数超过上限，请缩小日期范围或减少分组维度。",
              "VALIDATION_FAILED": "请求参数未通过校验，errors 列出具体字段。"
            }
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Error",
        "description": "RFC 9457 Problem Details（application/problem+json），所有 4xx/5xx 统一使用。"
      },
...
      "FilterOptions": {
        "properties": {
          "media": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "type": "array",
            "title": "Media"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "type": "array",
            "title": "Products"
          },
          "agencies": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "type": "array",
            "title": "Agencies"
          },
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "type": "array",
            "title": "Accounts"
          },
          "operators": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "type": "array",
            "title": "Operators"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "media",
          "products",
          "agencies",
          "accounts",
          "operators"
        ],
        "title": "FilterOptions"
      },
...
      "Option": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "value",
          "label"
        ],
        "title": "Option"
      },
...
      "RoiAnomalyResult": {
        "properties": {
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RoiAnomaly"
            },
            "type": "array",
            "title": "Items"
          },
          "data_as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data As Of"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rule",
          "items",
          "data_as_of"
        ],
        "title": "RoiAnomalyResult"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "本机 Java 业务服务",
      "x-environment": "development"
    },
    {
      "url": "http://127.0.0.1:18081",
      "description": "本机 Java 隔离验收示例，按需启动，非常驻入口",
      "x-environment": "testing"
    },
    {
      "url": "https://iaa-ops.invalid",
      "description": "公开演示规划占位，尚未部署；发布后替换为真实地址",
      "x-environment": "production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}